Bugfix: avoid fatal error when renaming chromosome X files. (#63)
* Bugfix: avoid fatal error when renaming chromosome X files.

The current imputation workflow script renames chromosome X files to ensure the PAR and non-PAR regions are properly ordered when sorted by lexicographical order in the compression workflow script.
This is achieved by first converting the filepath to a string, then substituting part of the filename, then calling `file(...)` to get a path again, and finally calling file.renameTo(...).

The current approach fails when the original path is an S3 path for two reasons:
* The path gets internally represented as a nextflow.cloud.aws.nio.S3Path instance, and loses the s3: protocol prefix when converted to string. Calling file(...) on the converted string results in an absolute path local to the main server.
* file.renameTo(...) does not support copying from S3 to a local path, so it fails immediately. Naive substitution for file.moveTo(...) causes a disk leak, since all Minimac 4 chromosome X outputs get copied to a local path outside the local workspace.

The solution proposed here is to simply avoid renaming files altogether, and instead use a content-aware sorting function in the compression step to ensure the files are submitted to bcftools in the correct order.

* Revert removing 'def' in compression_encryption_vcf.